requirements.txt
# Core Dependencies
customtkinter==5.2.0          # Modern GUI framework
pillow==10.1.0                # Image processing
pytesseract==0.3.10           # OCR engine wrapper
pdfplumber==0.10.3            # PDF text extraction
spacy==3.7.2                  # NLP processing
scikit-learn==1.3.2           # K-Means clustering
requests==2.31.0              # HTTP API calls
plotly==5.18.0                # Interactive charts
kaleido==0.2.1                # Static image export for plotly
pandas==2.1.4  

# Optional but Recommended
python-dotenv==1.0.0          # Environment variables
tqdm==4.66.1                  # Progress bars
loguru==0.7.2                 # Better logging